+++ Tests/KeyboardKitTests/Keyboard/KeyboardInputViewControllerTests.swift
@@ -210,7 +210,6 @@ class KeyboardInputViewControllerTests: QuickSpec {
 private class TestClass: KeyboardInputViewController, Mockable {
     
     lazy var viewWillSyncWithTextDocumentProxyRef = MockReference(viewWillSyncWithTextDocumentProxy)
-    lazy var setupKeyboardRef = MockReference(setupKeyboard)
     lazy var performAutocompleteRef = MockReference(performAutocomplete)
     lazy var resetAutocompleteRef = MockReference(resetAutocomplete)
     
@@ -232,10 +231,6 @@ private class TestClass: KeyboardInputViewController, Mockable {
     override func viewWillSyncWithTextDocumentProxy() {
         super.viewWillSyncWithTextDocumentProxy()
         mock.invoke(viewWillSyncWithTextDocumentProxyRef, args: ())
-    }
-    
-    override func setupKeyboard() {
-        mock.invoke(setupKeyboardRef, args: ())
     }
     
     override func performAutocomplete() {